pyproject.toml
[tool.poetry]
name = "discord-auth"
version = "0.1.0"
description = "Discord authentication service"
authors = ["Ian P <ian@aotc.us>"]
repository = "https://github.com/GoonAuthNetwork/discord-auth"
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "0.17.0"
pydantic = "^1.8.0"
loguru = "0.5.3"
httpx = "^0.16.1"
dispike = "^1.0.1-beta.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
black = "^20.8b1"
pytest = "^6.2.4"
pytest-asyncio = "^0.15.1"
poethepoet = "^0.10.0"

[tool.poe.env]
development="True"

[tool.poe.tasks]
start={ script="scripts:start"}
create_commands={ script="scripts:create_commands" }
delete_commands={ script="scripts:delete_commands" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
